+++ synquill_gen/CHANGELOG.md
@@ -1,8 +1,11 @@
-## 0.5.1
+## 0.7.0
+- Generate implementation for `getAllExcludingIds` in DAOs.
+- Explicitly add `@override` to `saveModel`, `deleteById`, and `deleteAll` in generated code.
+- Synchronized version with `synquill`.
 
+## 0.5.1
 - Generate `localOnly` property for repository implementations.
 > Note: code must be re-generated via build_runner
 
 ## 0.5.0
-
 - Initial version.